Fix filebrowser script escaping
We must not entity encode inline JS scripts, but rather make sure that
there is no plain `</script` contained to avoid that script parsing
ends prematurely.

Well, actually, we should not use inline scripts at all, but that is
what we have to, at least assuming there is still a usable external
filebrowser.