Define functions in init.php or admin.php
Misusing required_classes.php for this barely makes sense.
`codeeditor_getThemes()` is only required for the plugin configuration,
so it's perfectly fine to define it in admin.php.  Defining the
interface functions in required_classes.php is a minor convenience,
although having them in init.php is again perfectly fine, since clients
are supposed to include that file anyway[1].

[1] <https://wiki.cmsimple-xh.org/archiv/doku.php/plugin_interfaces#the_plugin_author>